+++ packages/expo-crypto/ios/CryptoModule.swift
@@ -10,7 +10,21 @@ public class CryptoModule: Module {
     AsyncFunction("digestStringAsync", digestString)
 
     Function("digestString", digestString)
+
+    AsyncFunction("getRandomBase64StringAsync", getRandomBase64String)
+
+    Function("getRandomBase64String", getRandomBase64String)
   }
+}
+
+private func getRandomBase64String(length: Int) throws -> String {
+  var bytes = [UInt8](repeating: 0, count: length)
+  let status = SecRandomCopyBytes(kSecRandomDefault, length, &bytes)
+
+  guard status == errSecSuccess else {
+    throw FailedGeneratingRandomBytesException(status)
+  }
+  return Data(bytes).base64EncodedString()
 }
 
 private func digestString(algorithm: DigestAlgorithm, str: String, options: DigestOptions) throws -> String {
@@ -36,5 +50,11 @@ private func digestString(algorithm: DigestAlgorithm, str: String, options: Dige
 private class LossyConversionException: Exception {
   override var reason: String {
     "Unable to convert given string without losing some information"
+  }
+}
+
+private class FailedGeneratingRandomBytesException: GenericException<OSStatus> {
+  override var reason: String {
+    "Generating random bytes has failed with OSStatus code: \(param)"
   }
 }