+++ packages/expo-crypto/src/ExpoCrypto.web.ts
@@ -27,6 +27,16 @@ export default {
     }
     throw new CodedError('ERR_CRYPTO_DIGEST', 'Invalid encoding type provided.');
   },
+  getRandomBytes(length: number): Uint8Array {
+    const array = new Uint8Array(length);
+    // @ts-ignore
+    return (window.crypto ?? window.msCrypto).getRandomValues(array);
+  },
+  async getRandomBytesAsync(length: number): Promise<Uint8Array> {
+    const array = new Uint8Array(length);
+    // @ts-ignore
+    return (window.crypto ?? window.msCrypto).getRandomValues(array);
+  },
 };
 
 function hexString(buffer: ArrayBuffer): string {