+++ packages/expo-random/src/Random.ts
@@ -3,7 +3,14 @@ import { UnavailabilityError } from 'expo-modules-core';
 
 import ExpoRandom from './ExpoRandom';
 
+const warnIsDeprecated = (functionName: string) =>
+  console.warn(
+    `expo-random is deprecated in favor of expo-crypto: use ExpoCrypto.${functionName}()instead. https://docs.expo.dev/versions/latest/sdk/crypto/`
+  );
+
 function assertByteCount(value: any, methodName: string): void {
+  warnIsDeprecated('assertByteCount');
+
   if (
     typeof value !== 'number' ||
     isNaN(value) ||
@@ -25,6 +32,7 @@ function assertByteCount(value: any, methodName: string): void {
  * @return An array of random bytes with the same length as the `byteCount`.
  */
 export function getRandomBytes(byteCount: number): Uint8Array {
+  warnIsDeprecated('getRandomBytes');
   assertByteCount(byteCount, 'getRandomBytes');
   const validByteCount = Math.floor(byteCount);
   if (__DEV__) {
@@ -55,6 +63,7 @@ export function getRandomBytes(byteCount: number): Uint8Array {
  * @return A promise that fulfills with an array of random bytes with the same length as the `byteCount`.
  */
 export async function getRandomBytesAsync(byteCount: number): Promise<Uint8Array> {
+  warnIsDeprecated('getRandomBytesAsync');
   assertByteCount(byteCount, 'getRandomBytesAsync');
   const validByteCount = Math.floor(byteCount);
   if (ExpoRandom.getRandomBytesAsync) {